Hide FMT better from interface
